Increase accuracy of geocoded cities
Prior to this commit, the latitude and longitude for most of the cities
was wildly inaccurate and in some cases, placing cities on opposite
sides of the country.  This commit refines those locations by using the
same geocoder but excluding ` city` from `place` and a mostly
comprehensive visual inspection of the resulting placements.