Fix redundancy between Enter and Update in 25_adding_values.html
I think there is a bit of redundancy in these two code sections?  In the `bars.enter()` block `"x"`, `"y"`, `"width"`, `"height"`, and `"fill"` are set.  In the `bars.transition()` block, the only things that change are the `"x"` and `"width"`, so I think including `"y"` and `"height"` there is unnecessary.